requirements.txt
django
djangorestframework
drf_yasg
black
pytest-django
pytest
django-rest-framework-nested
djangorestframework-simplejwt
django-cors-headers
pillow
psycopg2-binary
djangorestframework-bulk
drf-nested-routers
djoser
celery